make it possible to re-upload CSVs and re-run selection
The problem was various data structures being modified during
initialisation when the code assumed it wasn't.  So now we keep
extra copies of the original data around so we can re-do the
modifications from the original starting point.

Also, when the people have already been uploaded, if the categories
are re-uploaded we need to re-run the code that combines the
categories and people data.