Write default settings file if not present
and show a message to the user so they know where to find the
settings file.  The default settings are in a string in
stratification.py

As part of this I've deleted the separate settings file in the
repo so that we don't end up editing one copy of the settings
but not the other.