Implement LexiMin algorithm, a refined version of MaxiMin that maximizes not only the lowest probability, but breaks ties in favor of the second-lowest, further ties in favor of the third-lowest and so forth. The computation requires the commercial optimization library Gurobi, which is more reliable and faster and exposes more levers.
Activate LexiMin by default, but fall back to MaxiMin if Gurobi is not
available.

Remove `fair_to_households` configuration option, which was to our
knowledge not used and hard to maintain.

Improve which quota relaxations are suggested when quotas are
infeasible. The new formultation will be more reluctant to make big
relaxations for large groups, but is still reluctant to lower the quotas
of small groups.

In all algorithms except Legacy, write the distribution over panels to
the file table.csv for manual drawing of the panel.

Input files can now contain special characters in UTF-8 encoding.