Re-add tests and fix bug in how we use mip model
The bug in stratification.py appears to happen due to the mip model
weights all being set to 1.  The tests say that 46 results are desired,
but every time the first loop (in `_generate_initial_committees()`)
actually generates 49 results.  I fixed it by setting the initial results
to a random number between 0.8 and 1.0.

I also tried just setting every initial weight to 0.9, 0.99 etc, but we
still got 49 members in the committee on the first loop.

The tests were taken from those deleted in commit
https://github.com/sortitionfoundation/stratification-app/commit/9cbb09d44257d4fdbc36e96daf19f7e3f521d1ac

The tests have been reformatted, imports updated and the expected error
changed in two tests from InfeasibleQuotasError to InfeasibleQuotasCantRelaxError.