Make the tests for the legacy mode pass
By fixing a bug in the code that deletes people after a selection.  It
would try to look up all the categories of the person in the target
categories dict - but for `first_name` and the like there is no match
and so the code was throwing an exception at that point.  Now we check
if there is a matching category before trying to use it.  Non-matching
categories are skipped.