Slightly improve messages in central box
Rather than having two messages appear about asking for data as soon as
we start, we instead just show one message, and show the people tab
message when we get there. Also have the messages show what data we're
asking for and which tab is being read to fetch the data.